refactor(cli): move the device commands out of core/cli.py
Extract `devices`, `identify` and the `_print_raw_port_debug` helper into the
new `modules/device` package. The bodies are moved verbatim; as in Fase 4 the
two commands go from `@cli.command()` to `@click.command()` and are registered
explicitly in `build_cli()`. `core/cli.py` goes from 570 to 465 lines.

`_print_raw_port_debug` is private to `devices`, so it travels with it rather
than moving down to `core/` -- no new cross-package import.

Those were the last users of `.catnip`, `.device_utils` and `.usb_connection`
in core/cli.py, so all three imports are dropped along with `rich.table` and
`rich.box`.

Verified: no diff against tests/snapshots/cli_tree_linux.txt, `find_packages`
lists modules.device, `catnip devices` still runs, and `pytest tests/ -q`
gives 328 passed / 3 failed -- the same preexisting failures.
