+++ catnip/modules/core/cli.py
@@ -13,12 +13,11 @@ import sys
 
 # Internal
 from ..utils._version import __version__
-from .device_utils import get_device_or_exit
-from .catnip import catnip_get_devices
-from .usb_connection import ShellConnection, CATSNIFFER_VID, CATSNIFFER_PID
 
 # Command groups assembled by build_cli()
 from ..sniff.cli import sniff as _sniff
+from ..device.cli import devices as _devices
+from ..device.cli import identify as _identify
 from ..firmware.cli import flash as _flash
 from ..firmware.cli import update as _update
 from ..firmware.cli import restore as _restore
@@ -32,8 +31,6 @@ from ..protocols.cli.vhci import vhci as _vhci
 import click
 from rich.logging import RichHandler
 from rich.panel import Panel
-from rich.table import Table
-from rich import box
 
 from ..utils.output import (
     console,
@@ -132,110 +129,6 @@ def cli(verbose):
     if verbose:
         logger.level = logging.INFO
     pass
-
-
-@cli.command()
-@click.option(
-    "--debug",
-    is_flag=True,
-    default=False,
-    help="Show raw USB port info for each interface (useful for diagnosing Windows port mapping).",
-)
-def devices(debug: bool) -> None:
-    """List connected CatSniffer devices"""
-    devs = catnip_get_devices()
-    if not devs:
-        print_warning("No CatSniffer devices found.")
-        if debug:
-            _print_raw_port_debug()
-        return
-
-    # Add a table to display devices
-    table = Table(title=f"Found {len(devs)} CatSniffer device(s)", box=box.ROUNDED)
-    table.add_column("Device", style=STYLES["device"], justify="left")
-    table.add_column("Cat-Bridge (CC1352)", style="cyan", justify="left")
-    table.add_column("Cat-LoRa (SX1262)", style="cyan", justify="left")
-    table.add_column("Cat-Shell (Config)", style="cyan", justify="left")
-
-    for dev in devs:
-        bridge_status = dev.bridge_port or "[red]Not found[/red]"
-        lora_status = dev.lora_port or "[red]Not found[/red]"
-        shell_status = dev.shell_port or "[red]Not found[/red]"
-
-        table.add_row(str(dev), bridge_status, lora_status, shell_status)
-
-    print_empty_line()
-    console.print(table)
-
-    if debug:
-        _print_raw_port_debug()
-
-
-def _print_raw_port_debug() -> None:
-    """Print raw pyserial port info for all CatSniffer interfaces."""
-    from serial.tools import list_ports
-
-    cat_ports = [
-        p
-        for p in list_ports.comports()
-        if p.vid == CATSNIFFER_VID and p.pid == CATSNIFFER_PID
-    ]
-
-    if not cat_ports:
-        console.print("[red]No CatSniffer USB interfaces visible to pyserial.[/red]")
-        return
-
-    raw = Table(title="Raw USB port info (debug)", box=box.SIMPLE)
-    raw.add_column("Port", style="cyan")
-    raw.add_column("Description")
-    raw.add_column("HWID")
-    raw.add_column("Location")
-    raw.add_column("Interface")
-    raw.add_column("Serial#")
-
-    for p in sorted(cat_ports, key=lambda x: x.device):
-        raw.add_row(
-            p.device,
-            p.description or "",
-            p.hwid or "",
-            p.location or "",
-            getattr(p, "interface", None) or "",
-            p.serial_number or "",
-        )
-
-    console.print(raw)
-
-
-@cli.command()
-@click.option(
-    "--device",
-    "-d",
-    default=None,
-    type=int,
-    help="Device ID (for multiple CatSniffers)",
-)
-def identify(device) -> None:
-    """Send identification command to CatSniffer device"""
-    dev = get_device_or_exit(device)
-
-    if not dev.shell_port:
-        print_error("Shell port not available for this device!")
-        exit(1)
-
-    print_info(f"Sending 'Identify' command to {dev} on port {dev.shell_port}...")
-
-    try:
-        shell = ShellConnection(port=dev.shell_port, timeout=1.0)
-        with shell:
-            response = shell.send_command("identify", timeout=1.0)
-            if response:
-                print_info(f"Response: {response}")
-
-        print_success("Identification command sent successfully!")
-
-    except Exception as e:
-        print_error(f"Failed to send identification command: {str(e)}")
-        exit(1)
 
 
 # ===================== Shell Completion Commands =====================
@@ -550,6 +443,8 @@ def build_cli() -> click.Group:
     cli.add_command(_sniff)
     cli.add_command(_cativity)
     cli.add_command(_meshtastic)
+    cli.add_command(_devices)
+    cli.add_command(_identify)
     cli.add_command(_flash)
     cli.add_command(_update)
     cli.add_command(_restore)
