+++ catnip/modules/device/__init__.py
@@ -1,0 +1,5 @@
+"""``catnip devices|identify`` - device discovery and identification.
+
+Deliberately empty: ``modules/core/cli.py`` imports ``modules.device.cli``
+statically, so anything eager here would be paid on every ``catnip --help``.
+"""
